KeggModuleEnrichment in kegg.py -> metabolism/encrichment.py
